+++ rbx_reflector/src/defaults.rs
@@ -62,8 +62,8 @@ fn apply_instance_defaults(database: &mut ReflectionDatabase, instance: &Instanc
         let property_name = Cow::Owned(property_name.clone());
 
         match property_value.ty() {
-            // We skip these types because their defaults aren't useful.
-            VariantType::Ref | VariantType::SharedString => continue,
+            // We skip the Ref type because its default value is not useful.
+            VariantType::Ref => continue,
 
             _ => class
                 .default_properties